orchestrator/pyproject.toml
[project]
name = "surveillance-orchestrator"
version = "0.1.0"
description = "Orchestrator for surveillance data pipeline"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-storage>=2.14.0",
    "google-cloud-bigquery>=3.17.0",
    "dbt-core>=1.7.0",
    "dbt-bigquery>=1.7.0",
    "pyyaml>=6.0",
    "structlog>=24.1.0",
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-exporter-otlp>=1.22.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
